// lib/terfy.d.ts
export type TerfyState = ReturnType<typeof rootReducer>;
// ...
export type TerfyDispatch = ReturnType<typeof configureStore>['dispatch'];
// ...
export type ITab = {
  uid: string;
  title: string;
  isActive: boolean;
  hasActivity: boolean;
};